pyproject.toml
[tool.poetry]
name = "market-ts-ml"
version = "0.1.0"
description = "Loading, preprocessing and forecasting stock market time series data"
authors = ["Konstantin Ustinov <stburnv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
numpy = "^1.26.4"
python-binance = "^1.0.19"
pybit = "^5.8.0"
python-dotenv = "^1.0.1"
yfinance = "^0.2.40"
scikit-learn = "^1.5.1"
matplotlib = "^3.9.1"
plotly = "5.22.0"
statsmodels = "^0.14.2"
ipykernel = "^6.29.5"
nbformat = "^5.10.4"
seaborn = "^0.13.2"
optuna = "^3.6.1"
shap = "^0.46.0"
catboost = "^1.2.5"
optuna-dashboard = "^0.15.1"
aiogram = "^3.10.0"
g4f = "^0.3.2.2"
curl-cffi = "^0.7.1"
torch = "^2.4.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
